fix: odd intermittent issue with load function (#21)
this issue is very confusing, but vim.iter causes a bug about 1 in 5
times with a new handler I was writing, but the previous version of this
function throws no errors. It also doesnt throw the error from my new
handler, but rather, from the event handler, and it mentions the
deferredUIEnter event?