fix: altered loading order for startup plugins
some minor order changes to require('lz.n').load

This change is mostly to prevent issues with people doing
less-adviseable things such as calling trigger_load from places.

This ensures startup plugins have the same environment that the lazy ones do.

Namely, the ability to call trigger_load with a name instead of only being able to call trigger_load with a plugin spec, due to the plugin not yet being in state.